+++ R/utils.recalc.freqhomsnp.r
@@ -27,14 +27,10 @@
 #' @author Custodian: Luis Mijangos (Post to
 #' \url{https://groups.google.com/d/forum/dartr})
 
-#' @seealso \code{utils.recalc.metrics} for recalculating all metrics,
-#' \code{utils.recalc.callrate} for recalculating CallRate,
-#' \code{utils.recalc.freqhomref} for recalculating frequency of homozygous
-#' reference, \code{utils.recalc.avgpic} for recalculating AvgPIC,
-#' \code{utils.recalc.freqhet} for recalculating frequency of heterozygotes,
-#' \code{gl.recalc.maf} for recalculating minor allele frequency,
-#' \code{gl.recalc.rdepth} for recalculating average read depth
+#' @seealso \code{\link{gl.recalc.metrics}} for recalculating all
+#' locus metrics
 
+#' @keywords internal
 # @export
 #' @return The modified genlight object.
 
@@ -53,10 +49,10 @@ utils.recalc.freqhomsnp <- function(x,
                      verbose = verbose)
     
     # CHECK DATATYPE
-    datatype <- utils.check.datatype(x, verbose = verbose)
+    datatype <- utils.check.datatype(x, accept = "SNP", verbose = verbose)
     
     # Check monomorphs have been removed up to date
-    if (x@other$loc.metrics.flags$monomorphs == FALSE) {
+    if (!isTRUE(x@other$loc.metrics.flags$monomorphs)) {
         if (verbose >= 2) {
             cat(
                 warn(
@@ -72,7 +68,7 @@ utils.recalc.freqhomsnp <- function(x,
     
     if (is.null(x@other$loc.metrics$FreqHomSnp)) {
         x@other$loc.metrics$FreqHomSnp <- array(NA, nLoc(x))
-        if (verbose >= 2) {
+        if (verbose >= 3) {
             cat(
                 report(
                     "  Locus metric FreqHomSnp does not exist, creating slot @other$loc.metrics$FreqHomSnp\n"